requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── LLMs ──────────────────────────────────────────────────────────────────────
google-generativeai>=0.7.0
langchain>=0.2.0
langchain-community>=0.2.0

# ── OCR & document parsing ────────────────────────────────────────────────────
pytesseract>=0.3.10
Pillow>=10.0.0
pdfplumber>=0.11.0
pdf2image>=1.17.0
python-docx>=1.1.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
requests>=2.31.0
